refactor: update import paths for TaskModal and TaskCreateModal; remove unused components and files
